requirements.txt
streamlit
numpy
tensorflow
opencv-python-headless
Pillow
ultralytics
matplotlib
